Restore OreSat source files, revert Node ID
The Node IDs will be stored in flash on the nodes, and ORESAT_DEFAULT_ID will trigger a read of this ID and use of it. It shouldn't be baked into the firmware itself as the firmware should just use a generic image for multiple devices. The node id config option is for testing purposes mostly