Reduce gyro output rate/filter bandwidth and sample range
The satellite isn't expected to approach 2000deg/s, that would be 300+
RPM, so Bad Things have happened if it is getting close to that.
Dropping it down to 1000deg/s means we get an extra bit at the lower end
of the scale.

Additionally the output from the gyro was pretty noisy just sitting
there stationary on a desk, on the order of +-2deg/s. The table might be
vibrating, but out sample rate is only 1Hz so we can turn the output
data rate (ODR) and thus the filter bandwidth all the way down to
100Hz and 32Hz respectively (the minimum setting). This made the noise
drop way down, to under 1deg/s or about 2LSB/s.