Updated oresat-gen-fw-files invocation
And also updated the oresat-configs install instructions.

In the latest version of oresat-configs the invocation for
oresat-gen-fw-files changed slightly, I didn't realize it was being used
as part of the build scripts here. This fixes up these build scripts to
use the new invocation, which is changing the argument `oresat0.5` to
`--oresat 0.5`.