Add board/firmware versions to OD generator
This patch ensures that the board and firmware versions make it
somewhere into the built artefacts. The script itself puts them into the
object dictionary at the appropriate locations.

The firmware version is pretty easy, it's just the output of `git
describe`. As long as tags are maintained it will do the right thing.

Board version is kind of nebulous, currently it just pulls the name of
the board config directory in. There's not really a better string
defined anywhere that I can see? Also because of the way the OD is
generated it's hard to get something from inside a file like a #define
so using the directory is easy to implement.