Clean up github action, fix git describe breaking CI
With github actions it only checks out a shallow copy of the repo, not
including tags. git describe by default looks for tags, but the --always
flag added here lets it fall back to an acceptable default when it
doesn't find a tag.

While I was doing that I cleaned up the current test workflow:
- latest versions of all the actions
- oresat-configs now can come straight from pypi, we don't need to build
  it ourselves
- trim packages and tools to only what's used for the compiles below.
  There's a lot less packages to download now and it's noticeably
  faster.