Add SOLAR_V5 board
Now that programmed board names are pulled from the board directory, the
solar firmware should have a correct name. The current flying hardware
is Solar v5.1/v5.3 and v4s "don't exist". From the firmware's point of
view and for the actual configs, the hardware is identical and SOLAR_V5
is a copy of SOLAR_V4.