Toolchain improvements for GDB, VulCAN
This replaces the outdated arm-none-eabi-gdb with gdb-multiarch, which
modern distros package instead. This also adds a udev rule and systemd
service which will automatically create a vulan0 network device when the
adapter is plugged in. Place 99-vulcan.rule in /etc/udev/rules.d and
slcand@.service in /etc/systemd/system. The rules for both will then
have to be reloaded: `udevadm control --reload` and `systemctl
daemon-reload`.