Overhaul F0 bootloader
This overhauls the F0 bootloader and associated tooling. The original
bootloader's command can IDs were incompatible with CANopen's address
assignment and so the instigator for this patch was to reassign them to
the reserved unused addresses in CANopen's protocol. While I was there
I've fixed a number of other issues too:
- Fixed linker scripts and memory layout.
- Fixed memory region addressing, using linker defined symbols instead
  of ints.
- Improved compatibility with the protocol that this bootloader is based
  on, AN3154
- Fixed a number of buffer overflows and off-by-one errors in the
  protocol
- Rearranged the metadata section to not have 1kb of empty space in
  every image. This speeds up bootloader image write time.
- Enabled CRC hardware acceleration.
- Added a python based unit test suite
- Overhauled the python write script to be used as a library as well
- Fixed type issues
- Reduced oresat common dependencies