Rework openocd/gdb configuration
Follwing from 98174ea5f23eb8c which updated the transport for the nucleo
board, my initial thought was to update it for the rest of the boards
but that got me thinking why does the board need to configure the
debugger? Surely I could use any debugger with a specific board (not
considering literally the nucleo that has a built-in debugger). The only
thing the board needs to self-configure is the MCU target.

This splits the old oocd.cfg into a -target and -interface variant where
the targets are unique to the board but only one interface exists, in
the toolchain directory. Additionally the gdboocd.cmd was made target
agnostic via $_TARGETNAME and subsequently moved to the toolchain dir as
well.

Finally the actual debugger interface was updated to stlink-dap which
appears to be the proper one for modern stlinks with the dapdirect_swd
transport.