Rewrite flash_node_id to use openocd builtin options_write
It turns out openocd knows how to write the options bytes without us
doing it manually by hand. This both handles more edge cases and is much
quicker than us doing it by hand. The quickness as it turns out matters
a lot, given the shenanigans programming the solar cards via trying to
hold a ribbon cable to bare contacts. It was difficult to get a
connection stable for longer than a few hundred milliseconds.

Now that openocd handles all the details too we don't need to have a
detailed log of how to set the option bytes.