Fix gitignore excluding gdboocd.cmd
I'm a big fan of gitignores only excluding the things that a project
actually produces. The proliferation of bulk template ignores tends to
cause hard to diagnose git issues, as here where *.cmd files were
ignored as part of ignoring linux kernel build files but we use them
for the suffix of gdb command files.

This file was supposed to go in the last patch.