MAX17205: Improve documentation, fix integer overflows
The bulk of this is annotation the return unit for each of the
max17205Read* functions and adding references to the datasheet where all
the conversion values come from. In addition the following fixes were
made:
- max17205ReadMaxMinVoltage return values changed from uint8_t to
  uint16_t as the max value could overflow.
- max17205ReadTime return value changed from uint16_t to uint32_t as the
  max value could overflow. The registers that we read using this
  function the datasheet implies that they probably won't overflow a
  uint16_t in seconds so the OD types have been left as is but clamped
  as a safety measure.
- max17205ReadMaxMinCurrent conversion was off by a factor of 10,
  40_000 -> 400_000.
- In worker.c stop_worker could return an uninitialized value, but the
  compiler would only warn about this with LTO off.

Also max17205ReadBattVoltage was renamed max17205ReadBatt as it is only
valid reading the BATT register, the register address was inlined.