Fix overflows in max17205 driver
It looks like the types were selected based on the expected values of
the registers instead of the absolute minimum and maximum values. This
expands the types so that they can represent the full range, which is
helpful when printing debugging info, but then clamps the values when
writing to the Object Dictionary because in normal operation we
shouldn't see the extreme values.