Switch from continuous to triggered conversions
The read_avg_power_and_voltage function will now trigger a read and then
wait for it to be finished before updating the state. This will make
sure that all the readings are synchronized. In order to do this, the
mode must be set to triggered conversions when calling ina226Start.