Remove unecessary state member variables
As far as I can tell the NextDirection enum and the direction_up_flag
are only used in one place. Because of this, we should be able to remove
these variables and clean up a lot of the wordiness here. Also, I didn't
see anywhere where the voltage_threshold_flag variable was used so I
removed it. Lastly, when the slope_error is zero, previously the
function would multiply VREF_NEGATIVE by 0 and then return 0. I added a
final else branch for if the slope_error is 0.