Rename adcs to mag and add reaction-wheel-app
The ADCS card will now be called the mag card.
As a result, the oresat-adcs-app is renamed
the oresat-mag-app, as is the corresponding
folder in the apps folder.

A separate board for the reaction wheels will
be attached to the mag card and have its own
processor. So there is now a separate
oresat-reaction-wheel-app for that.